Isolate real UIState tests from process-global mock.module leaks
Split UIState class into ui-impl.ts so tests needing the real class
bypass mock.module. Fix prose/editor.test.ts to mock ../api directly
instead of global.fetch (which gets overridden by canvas-sync mocks).
